release 2.8.13: feed detail UI cleanup
- Always-visible title edit pencil (was hover-only, untappable on touch)
- Tags section collapsed by default
- Per-feed controls moved into a collapsed 'Feed settings' section,
  extracted to FeedSettingsPanel.tsx to match the other feed panels
