+++ frontend/src/pages/FeedDetail.tsx
@@ -2,17 +2,16 @@ import { useState } from 'react';
 import { Pencil } from 'lucide-react';
 import { useParams, Link } from 'react-router-dom';
 import { useQuery, useMutation, useQueryClient } from '@tanstack/react-query';
-import { getFeed, getEpisodes, refreshFeed, updateFeed, getNetworks, reprocessAllEpisodes, ReprocessAllResult, bulkEpisodeAction, BulkAction, UpdateFeedPayload } from '../api/feeds';
+import { getFeed, getEpisodes, refreshFeed, updateFeed, reprocessAllEpisodes, ReprocessAllResult, bulkEpisodeAction, BulkAction, UpdateFeedPayload } from '../api/feeds';
 import type { BulkActionResult } from '../api/types';
 import Artwork from '../components/Artwork';
 import CopyButton from '../components/CopyButton';
 import DropdownMenu from '../components/DropdownMenu';
 import EpisodeList from '../components/EpisodeList';
 import LoadingSpinner from '../components/LoadingSpinner';
-import TriStateSelect from '../components/TriStateSelect';
 import { FeedTagsEditor } from '../components/FeedTagsEditor';
-import { WHISPER_LANGUAGES, labelForLanguage } from '../utils/whisperLanguages';
 import { feedDisplayTitle } from '../utils/feedTitle';
+import FeedSettingsPanel from './feeds/FeedSettingsPanel';
 import PodcastAdDistributionPanel from './feeds/PodcastAdDistributionPanel';
 import { formatStorage } from './settings/settingsUtils';
 import { stripHtml } from '../utils/stripHtml';
@@ -38,16 +37,11 @@ function reprocessModeVerb(mode: string): string {
 function FeedDetail() {
   const { slug } = useParams<{ slug: string }>();
   const queryClient = useQueryClient();
-  const [isEditingNetwork, setIsEditingNetwork] = useState(false);
   const [isEditingTitle, setIsEditingTitle] = useState(false);
   const [editTitle, setEditTitle] = useState('');
   const [showReprocessConfirm, setShowReprocessConfirm] = useState(false);
   const [selectedReprocessMode, setSelectedReprocessMode] = useState<'reprocess' | 'full' | 'llm'>('reprocess');
   const [reprocessResult, setReprocessResult] = useState<ReprocessAllResult | null>(null);
-  const [editNetworkOverride, setEditNetworkOverride] = useState<string>('');
-  const [editDaiPlatform, setEditDaiPlatform] = useState('');
-  const [editAutoProcessOverride, setEditAutoProcessOverride] = useState<string>('global');
-  const [editMaxEpisodes, setEditMaxEpisodes] = useState<string>('');
 
   // Pagination state
   const [page, setPage] = useState(1);
@@ -83,11 +77,6 @@ function FeedDetail() {
   const totalEpisodes = episodesData?.total ?? 0;
   const totalPages = Math.max(1, Math.ceil(totalEpisodes / pageSize));
 
-  const { data: networks } = useQuery({
-    queryKey: ['networks'],
-    queryFn: getNetworks,
-  });
-
   const refreshMutation = useMutation({
     mutationFn: (opts?: { force?: boolean }) => refreshFeed(slug!, opts),
     onSuccess: () => {
@@ -100,7 +89,6 @@ function FeedDetail() {
     mutationFn: (data: UpdateFeedPayload) => updateFeed(slug!, data),
     onSuccess: () => {
       queryClient.invalidateQueries({ queryKey: ['feed', slug] });
-      setIsEditingNetwork(false);
       setIsEditingTitle(false);
     },
   });
@@ -140,39 +128,6 @@ function FeedDetail() {
   const saveTitleEdit = () => {
     updateMutation.mutate({ titleOverride: editTitle.trim() || null });
   };
-
-  const startEditingNetwork = () => {
-    setEditNetworkOverride(feed?.networkIdOverride || '');
-    setEditDaiPlatform(feed?.daiPlatform || '');
-    if (feed?.autoProcessOverride === true) {
-      setEditAutoProcessOverride('enable');
-    } else if (feed?.autoProcessOverride === false) {
-      setEditAutoProcessOverride('disable');
-    } else {
-      setEditAutoProcessOverride('global');
-    }
-    setEditMaxEpisodes(feed?.maxEpisodes ? String(feed.maxEpisodes) : '');
-    setIsEditingNetwork(true);
-  };
-
-  const saveNetworkEdit = () => {
-    let autoProcessOverride: boolean | null = null;
-    if (editAutoProcessOverride === 'enable') {
-      autoProcessOverride = true;
-    } else if (editAutoProcessOverride === 'disable') {
-      autoProcessOverride = false;
-    }
-
-    const maxEp = editMaxEpisodes ? parseInt(editMaxEpisodes, 10) : null;
-
-    updateMutation.mutate({
-      networkIdOverride: editNetworkOverride || null,
-      daiPlatform: editDaiPlatform || undefined,
-      autoProcessOverride: autoProcessOverride,
-      maxEpisodes: maxEp !== null && !isNaN(maxEp) ? Math.max(10, Math.min(maxEp, 500)) : null,
-    });
-  };
-
 
   const handleToggleSelect = (id: string) => {
     setSelectedIds(prev => {
@@ -279,7 +234,7 @@ function FeedDetail() {
                 </p>
               </div>
             ) : (
-              <div className="group flex items-start gap-2">
+              <div className="flex items-start gap-2">
                 <h1 className="text-2xl font-bold text-foreground min-w-0 break-words">
                   {feedDisplayTitle(feed)}
                 </h1>
@@ -292,7 +247,7 @@ function FeedDetail() {
                   onClick={startEditingTitle}
                   aria-label="Edit feed title"
                   title="Edit feed title"
-                  className="mt-1.5 shrink-0 p-1 rounded text-muted-foreground opacity-0 group-hover:opacity-100 focus:opacity-100 hover:bg-accent transition-opacity"
+                  className="mt-1.5 shrink-0 p-1 rounded text-muted-foreground hover:text-foreground hover:bg-accent transition-colors"
                 >
                   <Pencil className="w-4 h-4" />
                 </button>
@@ -306,166 +261,6 @@ function FeedDetail() {
               {feed.lastRefreshed && (
                 <span>Updated {new Date(feed.lastRefreshed).toLocaleDateString()}</span>
               )}
-              <span>Feed cap: {feed.maxEpisodes || 300}</span>
-            </div>
-
-            {/* Network / DAI Platform info */}
-            <div className="mt-3 flex flex-wrap items-center gap-3 text-sm">
-              {isEditingNetwork ? (
-                <div className="space-y-2">
-                  <div className="flex items-center gap-2">
-                    <label className="text-muted-foreground text-sm w-16 shrink-0">Network:</label>
-                    <select
-                      value={editNetworkOverride}
-                      onChange={(e) => setEditNetworkOverride(e.target.value)}
-                      className="flex-1 min-w-0 px-2 py-1 text-sm bg-secondary border border-border rounded"
-                    >
-                      <option value="">Auto-detect</option>
-                      {networks?.map((network) => (
-                        <option key={network.id} value={network.id}>
-                          {network.name}
-                        </option>
-                      ))}
-                    </select>
-                  </div>
-                  <div className="flex items-center gap-2">
-                    <label className="text-muted-foreground text-sm w-16 shrink-0">DAI:</label>
-                    <input
-                      type="text"
-                      value={editDaiPlatform}
-                      onChange={(e) => setEditDaiPlatform(e.target.value)}
-                      placeholder="e.g., megaphone, acast"
-                      className="flex-1 min-w-0 px-2 py-1 text-sm bg-secondary border border-border rounded"
-                    />
-                  </div>
-                  <div className="flex items-center gap-2">
-                    <label className="text-muted-foreground text-sm w-16 shrink-0">Feed cap:</label>
-                    <input
-                      type="number"
-                      value={editMaxEpisodes}
-                      onChange={(e) => setEditMaxEpisodes(e.target.value)}
-                      placeholder="300"
-                      min={10}
-                      max={500}
-                      className="w-20 px-2 py-1 text-sm bg-secondary border border-border rounded"
-                    />
-                  </div>
-                  <div className="flex gap-2">
-                    <button
-                      onClick={saveNetworkEdit}
-                      disabled={updateMutation.isPending}
-                      className="px-2 py-1 text-xs bg-primary text-primary-foreground rounded hover:bg-primary/90 disabled:opacity-50"
-                    >
-                      {updateMutation.isPending ? 'Saving...' : 'Save'}
-                    </button>
-                    <button
-                      onClick={() => setIsEditingNetwork(false)}
-                      className="px-2 py-1 text-xs bg-muted text-muted-foreground rounded hover:bg-accent"
-                    >
-                      Cancel
-                    </button>
-                  </div>
-                </div>
-              ) : (
-                <div className="flex items-center gap-3 flex-wrap">
-                  {feed.networkId && (
-                    <span className={`px-2 py-0.5 rounded text-xs font-medium ${
-                      feed.networkIdOverride
-                        ? 'bg-orange-500/20 text-orange-600 dark:text-orange-400'
-                        : 'bg-green-500/20 text-green-600 dark:text-green-400'
-                    }`}>
-                      {feed.networkIdOverride ? 'Override' : 'Detected'}: {feed.networkId}
-                    </span>
-                  )}
-                  {feed.daiPlatform && (
-                    <span className="px-2 py-0.5 bg-purple-500/20 text-purple-600 dark:text-purple-400 rounded text-xs font-medium">
-                      DAI: {feed.daiPlatform}
-                    </span>
-                  )}
-                  <button
-                    onClick={startEditingNetwork}
-                    className="text-xs text-muted-foreground hover:text-foreground"
-                  >
-                    {feed.networkId || feed.daiPlatform ? 'Edit' : '+ Add Network'}
-                  </button>
-                </div>
-              )}
-            </div>
-
-            {/* Podcast Settings - Always visible */}
-            <div className="mt-4 space-y-3">
-
-              {/* Auto-Process Control */}
-              <div className="flex flex-col sm:flex-row sm:items-center gap-2 sm:gap-3 text-sm">
-                <span className="text-muted-foreground whitespace-nowrap">Auto-Process:</span>
-                <div className="flex items-center gap-2 flex-wrap">
-                  <TriStateSelect
-                    value={feed.autoProcessOverride}
-                    onChange={(next) => updateMutation.mutate({ autoProcessOverride: next })}
-                    disabled={updateMutation.isPending}
-                    className="px-2 py-1.5 text-sm bg-secondary border border-border rounded flex-1 sm:flex-none min-w-0"
-                  />
-                  {feed.autoProcessOverride !== null && feed.autoProcessOverride !== undefined && (
-                    <span className={`px-2 py-0.5 rounded text-xs font-medium ${
-                      feed.autoProcessOverride
-                        ? 'bg-green-500/20 text-green-600 dark:text-green-400'
-                        : 'bg-red-500/20 text-red-600 dark:text-red-400'
-                    }`}>
-                      {feed.autoProcessOverride ? 'Enabled' : 'Disabled'}
-                    </span>
-                  )}
-                </div>
-              </div>
-
-              {/* Per-feed transcription language override */}
-              <div className="flex flex-col sm:flex-row sm:items-center gap-2 sm:gap-3 text-sm">
-                <span className="text-muted-foreground whitespace-nowrap">Language:</span>
-                <div className="flex items-center gap-2 flex-wrap">
-                  <select
-                    value={feed.languageOverride ?? ''}
-                    onChange={(e) => {
-                      const v = e.target.value;
-                      updateMutation.mutate({ languageOverride: v === '' ? null : v });
-                    }}
-                    disabled={updateMutation.isPending}
-                    className="px-2 py-1.5 text-sm bg-secondary border border-border rounded flex-1 sm:flex-none min-w-0 disabled:opacity-50"
-                  >
-                    <option value="">Global default</option>
-                    <option value="auto">Auto-detect (multilingual)</option>
-                    {WHISPER_LANGUAGES.map((l) => (
-                      <option key={l.code} value={l.code}>
-                        {l.name} ({l.code})
-                      </option>
-                    ))}
-                  </select>
-                  {feed.languageOverride && (
-                    <span className="px-2 py-0.5 rounded text-xs font-medium bg-blue-500/20 text-blue-600 dark:text-blue-400">
-                      Override: {labelForLanguage(feed.languageOverride)}
-                    </span>
-                  )}
-                </div>
-              </div>
-
-              <div className="flex flex-col sm:flex-row sm:items-center gap-2 sm:gap-3 text-sm">
-                <span className="text-muted-foreground whitespace-nowrap">Hide unprocessed:</span>
-                <div className="flex items-center gap-2 flex-wrap">
-                  <TriStateSelect
-                    value={feed.onlyExposeProcessedEpisodes}
-                    onChange={(next) => updateMutation.mutate({ onlyExposeProcessedEpisodes: next })}
-                    disabled={updateMutation.isPending}
-                    className="px-2 py-1.5 text-sm bg-secondary border border-border rounded flex-1 sm:flex-none min-w-0"
-                  />
-                  {feed.onlyExposeProcessedEpisodes !== null && feed.onlyExposeProcessedEpisodes !== undefined && (
-                    <span className={`px-2 py-0.5 rounded text-xs font-medium ${
-                      feed.onlyExposeProcessedEpisodes
-                        ? 'bg-green-500/20 text-green-600 dark:text-green-400'
-                        : 'bg-red-500/20 text-red-600 dark:text-red-400'
-                    }`}>
-                      {feed.onlyExposeProcessedEpisodes ? 'Hiding' : 'Showing all'}
-                    </span>
-                  )}
-                </div>
-              </div>
             </div>
           </div>
         </div>
@@ -539,6 +334,8 @@ function FeedDetail() {
           </div>
         </div>
       </div>
+
+      {slug && <FeedSettingsPanel feed={feed} slug={slug} />}
 
       {slug && <FeedTagsEditor slug={slug} />}
