+++ openapi.yaml
@@ -12,7 +12,7 @@ info:
     - Monitor system status and trigger cleanup operations
     - Manage cross-episode ad patterns with network and podcast scope
     - Submit corrections to improve ad detection accuracy
-  version: 2.8.12
+  version: 2.8.13
   contact:
     name: MinusPod
   license:
